requirements.txt
PyYAML
pygit2
